renaming files for consistency & importing new landing pageS
