+++ client/src/pages/Initial/UzmaLanding/UzmaLanding.jsx
@@ -1,8 +1,8 @@
-import './LandingPageUzma.scss';
-import MainFroshLogo from '../../assets/logo/frosh-main-logo-with-bg.svg';
-import InstagramLogo from '../../assets/social/instagram_icon.png';
+import './UzmaLanding.scss';
+import MainFroshLogo from '../../../assets/logo/frosh-main-logo-with-bg.svg';
+import InstagramLogo from '../../../assets/social/instagram_icon.png';
 
-export const LandingPageUzma = () => {
+export const UzmaLanding = () => {
   return (
     <div className="landing-page-uzma-header">
       <div className="landing-page-uzma-header-text">
