+++ client/src/pages/Initial/UzmaLanding/UzmaLanding.scss
@@ -1,4 +1,4 @@
-@import '../../scssStyles/mixins';
+@import '../../../scssStyles/mixins';
 
 .landing-page-uzma-header {
   display: flex;
